+++ packages/starter/templates/react-typescript-webpack/src/services/greet-me.ts
@@ -7,7 +7,7 @@ import { type Context, type Serialization } from "@quatico/magellan-shared";
 // In order for Magellan to know that this code can be executed on the server, it must be annotated with a comment // @service() decorator.
 
 // @service()
-export const greetMe = async (name: string, _context?: Context, _serialization?: Serialization): Promise<string> => {
+export const greetMe = async (name: string, context?: Context, serialization?: Serialization): Promise<string> => {
     // This code must be executed on the server. In the browser, accessing process.arch causes an error.
     return `Hello ${name}! I'm Magellan running on "${typeof window === "undefined" ? `${process.arch}" server` : "browser"}!`;
 };